Commit de limpeza nos métodos Https

+++ src/main/java/br/com/pines/dev/controller/ProductController.java
@@ -4,7 +4,6 @@ package br.com.pines.dev.controller;
 import br.com.pines.dev.model.Product;
 import br.com.pines.dev.model.dto.ProductDto;
 import br.com.pines.dev.repository.ProductRepository;
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
@@ -12,7 +11,6 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
 
 
-import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
@@ -31,31 +29,31 @@ public class ProductController {
         List<ProductDto> productsDto = new ArrayList<>();
         if (name==null) {
             List<Product> products = productRepository.findAll();
-            for (Product product : products) {
-                productsDto.add(product.conversor(product));
-            }
+            products.forEach(product -> productsDto.add(product.conversor()));
         } else {
             List<Product> products = productRepository.findByName(name);
-            for (Product product: products) {
-                productsDto.add(product.conversor(product));
-            }
+            products.forEach(product -> productsDto.add(product.conversor()));
         }
         return productsDto;
     }
 
+    @PreAuthorize("hasRole('ROLE_ADMIN')")
     @GetMapping("/{id}")
     @Transactional
-    public ProductDto productById(@PathVariable Long id){
-        Product product = productRepository.getReferenceById(id);
-        ProductDto productDto = product.conversor(product);
-        return productDto;
+    public ResponseEntity<ProductDto> productById(@PathVariable Long id){
+        Optional<Product> product = productRepository.findById(id);
+        if (product.isPresent()){
+            ProductDto productDto = product.get().conversor();
+            return ResponseEntity.ok(productDto);
+        }
+        return ResponseEntity.notFound().build();
     }
 
     @PreAuthorize("hasAnyRole('ROLE_USER', 'ROLE_ADMIN')")
     @PostMapping
     @Transactional
     public ResponseEntity<ProductDto> registrationProduct(@RequestBody ProductDto productDto) {
-        Product product = productDto.conversor(productDto);
+        Product product = productDto.conversor();
         productRepository.save(product);
         return ResponseEntity.ok(productDto);
     }
@@ -63,22 +61,25 @@ public class ProductController {
     @PreAuthorize("hasAnyRole('ROLE_USER', 'ROLE_ADMIN')")
     @DeleteMapping("/{id}")
     @Transactional
-    public void deleteProduct(@PathVariable Long id){
-        productRepository.deleteById(id);
+    public ResponseEntity<Product> deleteProduct(@PathVariable Long id){
+        Optional<Product> product = productRepository.findById(id);
+        if (product.isPresent()){
+            productRepository.deleteById(id);
+            return ResponseEntity.ok().build();
+        }
+        return ResponseEntity.badRequest().build();
     }
 
     @PreAuthorize("hasAnyRole('ROLE_USER', 'ROLE_ADMIN')")
     @PutMapping("/{id}")
     @Transactional
-    public ResponseEntity<Product> updateProduct(@PathVariable Long id, @RequestBody Product p) {
-
+    public ResponseEntity<Product> updateProduct(@PathVariable Long id, @RequestBody Product product) {
         return productRepository.findById(id)
-                .map(m -> {
-            m.setName(p.getName());
-            m.setDescription(p.getDescription());
-            m.setPrice(p.getPrice());
-            Product product = productRepository.save(m);
-            return ResponseEntity.ok().body(product);
+                .map(updatedProduct -> {
+            updatedProduct.setName(product.getName());
+            updatedProduct.setDescription(product.getDescription());
+            updatedProduct.setPrice(product.getPrice());
+            return ResponseEntity.ok().body(productRepository.save(updatedProduct));
         }).orElse(ResponseEntity.notFound().build());
     }
 }
